Task.php refactor
