+++ .htaccess
@@ -1,0 +1,13 @@
+AllowOverride All
+AddCharset UTF-8 .php
+
+<IfModule mod_headers.c>
+    Header unset ETag
+    <filesMatch "\.*$">
+        Header set Cache-Control "no-store, no-cache, must-revalidate"
+        Header set Cache"max-age=0, post-check=0 pre-check=0, replace=false"
+        Header set Pragma "no-cache"
+    </filesMatch>
+</IfModule>
+
+FileETag None
